Классы для парсинга json разнесены по файлам, добавлена аннотация Keep
