+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/Actor.kt
@@ -1,8 +1,5 @@
 package com.alexanderpodkopaev.androidacademyproject.data
 
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
 data class Actor(
     val id: Int,
     val name: String,
