+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/ActorJsonModel.kt
@@ -1,8 +1,10 @@
 package com.alexanderpodkopaev.androidacademyproject.data
 
+import androidx.annotation.Keep
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+@Keep
 @Serializable
 data class ActorJsonModel(
     val id: Int,
@@ -16,3 +18,11 @@ fun ActorJsonModel.convertToModel(imageBaseUrl: String): Actor {
         id = this.id, name = this.name, picture = imageBaseUrl + "original" + this.picture
     )
 }
+
+@Keep
+@Serializable
+data class AllActorsData(
+    val id: Int,
+    @SerialName("cast")
+    val actors: List<ActorJsonModel>
+)
