+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/MovieJsonModel.kt
@@ -1,8 +1,10 @@
 package com.alexanderpodkopaev.androidacademyproject.data
 
+import androidx.annotation.Keep
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+@Keep
 @Serializable
 data class MovieJsonModel(
     val id: Int,
