+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/MoviesJsonModel.kt
@@ -1,5 +1,6 @@
 package com.alexanderpodkopaev.androidacademyproject.data
 
+import androidx.annotation.Keep
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -39,3 +40,13 @@ data class MoviesJsonModel(
         voteCount = this.voteCount
     )
 }*/
+
+@Keep
+@Serializable
+data class AllMoviesData(
+    val page: Int,
+    @SerialName("results")
+    val movies: List<MoviesJsonModel>,
+    val total_pages: Int,
+    val total_results: Int
+)
